requirements.txt
pytest==6.2.1
pytest-reportlog==0.1.2
allure-pytest==2.8.29
allure-python-commons==2.8.29
#pytest-rerunfailures==9.0
configparser==5.0.0
pyyaml==5.3.1
requests==2.25.1
urllib3==1.26.2
requests-toolbelt==0.9.1
simplejson==3.17.2
ruamel.yaml==0.16.10